+++ samila/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """Samila modules."""
 from .genimage import GenerativeImage
-from .params import Projection, Marker, Gateway, VALID_COLORS, SAMILA_VERSION
+from .params import Projection, GenerateMode, Marker, Gateway, VALID_COLORS, SAMILA_VERSION
 from .errors import samilaDataError, samilaGenerateError
 __version__ = SAMILA_VERSION