+++ otherfiles/version_check.py
@@ -22,6 +22,7 @@ PARAMS_ITEMS = ['SAMILA_VERSION = "{0}"']
 META_ITEMS = ['% set version = "{0}" %']
 ISSUE_TEMPLATE_ITEMS = ["- Samila {0}"]
 SECURITY_ITEMS = ["| {0}           | :white_check_mark: |", "| < {0}         | :x:                |"]
+DEMO_ITEMS = ["### Version : {0}"]
 
 FILES = {
     "setup.py": SETUP_ITEMS,
@@ -34,6 +35,7 @@ FILES = {
         ".github",
         "ISSUE_TEMPLATE",
         "bug_report.yml"): ISSUE_TEMPLATE_ITEMS,
+    os.path.join("examples", "demo.ipynb"): DEMO_ITEMS
 }
 
 TEST_NUMBER = len(FILES)